+++ src/main/java/com/eureka/mindbloom/member/service/MemberService.java
@@ -1,6 +1,7 @@
 package com.eureka.mindbloom.member.service;
 
 import com.eureka.mindbloom.member.domain.Member;
+import com.eureka.mindbloom.member.dto.GetMemberProfileResponse;
 import com.eureka.mindbloom.member.dto.SignUpRequest;
 import com.eureka.mindbloom.member.dto.SignUpResponse;
 import com.eureka.mindbloom.member.dto.UpdateMemberProfileRequest;
